Rework forall atomic headers for consistency with other tests.
