+++ test/functional/forall/atomic-view/test-forall-atomic-view-hip.cpp
@@ -5,15 +5,7 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-///
-/// Source file containing basic functional tests for atomic operations with forall and views.
-///
-
-#include "tests/test-forall-atomic-view.hpp"
-
-#include "RAJA_test-forall-execpol.hpp"
-
-#include "../test-forall-atomic-utils.hpp"
+#include "test-forall-atomic-view.hpp"
 
 #if defined(RAJA_ENABLE_HIP)
 using HipAtomicForallViewTypes =
