+++ test/functional/forall/atomic-view/test-forall-atomic-view-openmp.cpp
@@ -5,13 +5,7 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-///
-/// Source file containing basic functional tests for atomic operations with forall and views.
-///
-
-#include "tests/test-forall-atomic-view.hpp"
-
-#include "../test-forall-atomic-utils.hpp"
+#include "test-forall-atomic-view.hpp"
 
 #if defined(RAJA_ENABLE_OPENMP)
 using OmpAtomicForallViewTypes =
