+++ test/functional/forall/atomic/test-forall-atomic-basic-openmp.cpp
@@ -5,13 +5,7 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-///
-/// Source file containing basic functional tests for atomic operations with forall.
-///
-
-#include "tests/test-forall-atomic-basic.hpp"
-
-#include "../test-forall-atomic-utils.hpp"
+#include "test-forall-atomic-basic.hpp"
 
 #if defined(RAJA_ENABLE_OPENMP)
 using OmpAtomicForallBasicTypes =
